port evm tests
